src/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
black==24.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.4.3
cryptography==41.0.7
defusedxml==0.8.0rc2
Django==3.2.7
django-appconf==1.0.6
django-celery-email==3.0.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-coverage==1.2.4
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-extensions==3.2.3
django-nose==1.4.7
django-phonenumber-field==7.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==23.3.0
flake8==7.0.0
flower==2.0.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nose==1.3.7
oauthlib==3.2.2
packaging==23.2
pathspec==0.12.1
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==8.1.0
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
